+++ src/sphinxquery/sphinxquery.cpp
@@ -272,56 +272,3 @@ void DotDump (const XQNode_t * pNode)
 	sph::RenderBsonPlan ( sRes, bson::MakeHandle ( dPlan ), true );
 	printf ( "\nhttps://dreampuf.github.io/GraphvizOnline/#%s\n", UrlEncode ( CSphString{sRes} ).cstr() );
 }
-
-
-bool sphParseExtendedQuery ( XQQuery_t & tParsed, const char * sQuery, const CSphQuery * pQuery, const TokenizerRefPtr_c& pTokenizer, const CSphSchema * pSchema, const DictRefPtr_c& pDict, const CSphIndexSettings & tSettings, const CSphBitvec * pMorphFields )
-{
-	XQParser_t qp;
-	bool bRes = qp.Parse ( tParsed, sQuery, pQuery, pTokenizer, pSchema, pDict, tSettings, pMorphFields );
-
-#ifndef NDEBUG
-	if ( bRes && tParsed.m_pRoot )
-		tParsed.m_pRoot->Check ( true );
-#endif
-
-#if XQDEBUG
-	if ( bRes )
-	{
-		printf ( "\n--- query ---\n" );
-		printf ( "%s\n", sQuery );
-		xqDump ( tParsed.m_pRoot, 0 );
-		DotDump ( tParsed.m_pRoot );
-		printf ( "\n--- query reconstructed ---\n" );
-		printf ( "%s\n", sphReconstructNode ( tParsed.m_pRoot ).cstr());
-		printf ( "---\n" );
-	}
-#endif
-
-	// moved here from ranker creation
-	// as at that point term expansion could produce many terms from expanded term and this condition got failed
-	tParsed.m_bSingleWord = ( tParsed.m_pRoot && tParsed.m_pRoot->dChildren().IsEmpty() && tParsed.m_pRoot->dWords().GetLength()==1 );
-	tParsed.m_bEmpty = qp.m_bEmpty;
-
-	return bRes;
-}
-
-
-class QueryParserPlain_c : public QueryParser_i
-{
-public:
-	bool IsFullscan ( const XQQuery_t & tQuery ) const override { return false; }
-	bool ParseQuery ( XQQuery_t & tParsed, const char * sQuery, const CSphQuery * pQuery, TokenizerRefPtr_c pQueryTokenizer, TokenizerRefPtr_c pQueryTokenizerJson, const CSphSchema * pSchema, const DictRefPtr_c& pDict, const CSphIndexSettings & tSettings, const CSphBitvec * pMorphFields ) const override;
-	QueryParser_i * Clone() const final { return new QueryParserPlain_c; }
-};
-
-
-bool QueryParserPlain_c::ParseQuery ( XQQuery_t & tParsed, const char * sQuery, const CSphQuery * pQuery, TokenizerRefPtr_c pQueryTokenizer, TokenizerRefPtr_c, const CSphSchema * pSchema, const DictRefPtr_c& pDict, const CSphIndexSettings & tSettings, const CSphBitvec * pMorphFields ) const
-{
-	return sphParseExtendedQuery ( tParsed, sQuery, pQuery, pQueryTokenizer, pSchema, pDict, tSettings, pMorphFields );
-}
-
-
-std::unique_ptr<QueryParser_i> sphCreatePlainQueryParser()
-{
-	return std::make_unique<QueryParserPlain_c>();
-}
